+++ src/kernel/uart.c
@@ -4,15 +4,15 @@
 #include <common/stdlib.h>
 
 // Memory-Mapped I/O output
-void mmio_write(uint32_t reg, uint32_t data)
+void mmio_write(uint64_t reg, uint64_t data)
 {
-    *(volatile uint32_t*)reg = data;
+    *(volatile uint64_t*)reg = data;
 }
 
 // Memory-Mapped I/O input
-uint32_t mmio_read(uint32_t reg)
+uint32_t mmio_read(uint64_t reg)
 {
-    return *(volatile uint32_t*)reg;
+    return *(volatile uint64_t*)reg;
 }
 
 // Loop <delay> times in a way that the compiler won't optimize away
